requirements.txt
# Testing
pytest==8.2.2
pytest-cov==5.0.0
multiprocess==0.70.16
pytest-asyncio==0.23.7

# Linting and Formatting
flake8==7.1.0
isort==5.13.2

# Type Checking
mypy==1.10.0
types-requests==2.32.0.20240712

# Implementation
huggingface_hub==0.23.4
PyGithub==2.3.0
GitPython==3.1.43
requests==2.32.3
aiohttp==3.9.5